Modified training and testing file
Modified the plotting function to accept an optional filepath parameter, if not None, the plots are saved as png files instead of plotted at runtime.
Modified the train_model function in order to handle interrupt signals that can occur when training in SLURM environment, avoiding losing the progress. The weights and the training results are saved as they are when the interrupt signal arrives.